reimpl brushing config, allow no labels/X labels only/Y labels only/all labels
add % change in Y for selected range to the upper Y (horizontal) label

fix brush rendering to correctly position the brush area for moving, etc.

Note: there is a bug when the lower bound crosses the upper, the new lower bound
steps forward one interval, and the bounds lines/labels are incorrectly marked
as highlighted. Maybe an interval rounding error? Not sure.