begin package-ifying
  - convert elm.json to package format
  - add docs to modules

Note also public API change to Sparklines.lineConfig, remove the Time.Interval
param (not needed).

Also public API change to Timeseries.intervalExtent and intervalExtentMultiple,
both return a Maybe now for compatibility with elm-visualization extent
functions. (This function is not used externally by the library but it is part
of the public API.)