add X label formatting, fix brush highlight-selection for line charts
  - add labelX, labelY to config (replacing toString)
  - add default time labels given interval, using ryannhg/date-format
  - change the config API a bit for brush labels

Note for continuous time scales, it's not safe to assumeevery observation
is equally spaced by a time interval. You could have gaps, AKA uneven time
intervals between observations. Thus, we can't use the same strategy as for
band scale of rounding down or up to the nearest interval. Instead, in this
commit we look at the intervals between each timepoint in the series: for
the upper bound, does the given x value fall between a highlighted point and
the next point; and for the lower bound, does the given x fall between the
previous point and a highlighted point? I am not 100% sure why this formula
works, but it seems to in practice.

Also: fixed bug where lowest point in series could not be selected in line
chart brushes. This had to do with the example code applying the padding to
the brush area twice.